Revert @llvm.isnan intrinsic patchset.
Please refer to
https://lists.llvm.org/pipermail/llvm-dev/2021-September/152440.html
(and that whole thread.)

TLDR: the original patch had no prior RFC, yet it had some changes that
really need a proper RFC discussion. It won't be productive to discuss
such an RFC, once it's actually posted, while said patch is already
committed, because that introduces bias towards already-committed stuff,
and the tree is potentially in broken state meanwhile.

While the end result of discussion may lead back to the current design,
it may also not lead to the current design.

Therefore i take it upon myself
to revert the tree back to last known good state.

This reverts commit 21bfb5a59abc69d5c888ace90363a21745d5a69c.
This reverts commit 62af80844cc70fc1168298fe1019d99d0a824478.
This reverts commit 8d4eeaefd1bbf9b4dbe91f3aadb5a126b6607e72.
This reverts commit d838608f7c10b9ba0a5265866dcf54c15beda6a4.
This reverts commit aa565843118858ae0f10990230d5f40fa055d77e.
This reverts commit c483ca586f469ef846e8326f7253c2d94fb0fe18.
This reverts commit f319a29f8c9be6c9bddc9c17adf3ae2a586c77fd.
This reverts commit 673589cd3fde40fe0d09331ddc89852d1e6c9e41.
This reverts commit b9bfe59854da642f4c043156ae62e8b70aa4dcc4.